[FEAT] 피드 관련 검증 로직 구현 (#67)
* refactor: AuthorResponse 이름 변경

* refactor: AuthorResponse 이름 변경으로 인한 수정

* feat: 피드 엔티티 Cascade 옵션 추가

- 피드 삭제시 연관된 Likes 엔티티 삭제를 위한 Cascade 옵션 추가

* refactor: 검증을 위해 Member 정보 받도록 수정

* style: 코드 리포맷팅

* feat: 좋아요 관련 검증 로직 추가

- 해당 피드 작성자만 좋아요를 누른 멤버를 조회할 수 있도록 수정
- 좋아요를 누르지 않은 피드에 대해 좋아요 취소 요청을 보낼 경우 오류 발생

* feat: 피드 삭제관련 검증 추가

- 해당 피드 작성자만이 피드를 삭제할 수 있도록 수정

* feat: swagger 설정 @Parameter(hidden = true) 추가

+++ src/main/java/kakao/rebit/feed/service/LikesService.java
@@ -27,8 +27,14 @@ public class LikesService {
     }
 
     @Transactional(readOnly = true)
-    public Page<LikesMemberResponse> getLikesMembers(Long feedId, Pageable pageable) {
+    public Page<LikesMemberResponse> getLikesMembers(MemberResponse memberResponse, Long feedId,
+            Pageable pageable) {
+        Member member = memberService.findMemberByIdOrThrow(memberResponse.id());
         Feed feed = feedService.findFeedByIdOrThrow(feedId);
+
+        if (!feed.isWrittenBy(member)) {
+            throw new IllegalArgumentException("피드 작성자만 좋아요 누른 유저들을 확인할 수 있습니다.");
+        }
         return likesRepository.findAllByFeedWithMember(feed, pageable)
                 .map(this::toLikesMemberResponse);
     }
@@ -49,6 +55,10 @@ public class LikesService {
     public void deleteLikes(MemberResponse memberResponse, Long feedId) {
         Member member = memberService.findMemberByIdOrThrow(memberResponse.id());
         Feed feed = feedService.findFeedByIdOrThrow(feedId);
+
+        if (!likesRepository.existsByMemberAndFeed(member, feed)) {
+            throw new IllegalArgumentException("좋아요 내역이 없습니다.");
+        }
 
         likesRepository.deleteByMemberAndFeed(member, feed);
     }
